backend-api/requirements.txt
streamlit
langchain
langchain-openai
python-dotenv
yfinance
pandas
plotly
numpy
ta
langgraph
pydantic
fpdf2
secure-smtplib
email-validator
kaleido
fastapi
uvicorn
twilio
python-multipart